Moved dependencies to super pom, removed them from all project pom(s) (versions only).
The versions not updated to the latest, but if more than one version was used (in different modules),
I have unified to the newest (used) one.
There are 3 components that used different versions for sub-components.
Probably not ok, to cleanup when we update the versions.
(i.e. org.codehaus.jackson uses "1.4.3" for jackson-core-lgpl, jackson-mapper-lgpl
and "1.5.0" for jackson-core-asl, jackson-jaxrs, jackson-mapper-asl, jackson-xc)